requirements.txt
requests
yaspin
clint
tk
openpyxl
reportlab
pypdf
pdf2image
pillow